backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
python-dotenv==1.0.1
httpx==0.27.0
sqlalchemy==2.0.23
aiosqlite==0.19.0

